Refactoring methods in MapViewModel and adding a ProductCategory model
