+++ MonMaraicher/Assets.xcassets/AccentColor.colorset/Contents.json
@@ -7,7 +7,7 @@
           "alpha" : "1.000",
           "blue" : "0.247",
           "green" : "0.231",
-          "red" : "0.921"
+          "red" : "0.922"
         }
       },
       "idiom" : "universal"
