+++ MonMaraicher/Assets.xcassets/biscuit.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chou-rouge.png",
+      "filename" : "biscuit.png",
       "idiom" : "universal"
     }
   ],
