+++ MonMaraicher/Assets.xcassets/filter.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "chou-rouge.png",
+      "filename" : "filter.png",
       "idiom" : "universal"
     }
   ],
