+++ MonMaraicher/Model/Farmer/Farmers.swift
@@ -144,53 +144,3 @@ struct Certificats: Decodable, Hashable {
         case url
     }
 }
-
-enum ProductsImages: String, CaseIterable {
-    case artichoke = "artichaut"
-    case cherry = "cerise"
-    case chicory = "chicorée"
-    case cabbage = "choux"
-    case cauliflower = "choux-fleurs"
-    case beeswax = "cire d'abeille"
-    case spinach = "épinard"
-    case strawberry = "fraise"
-    case lettuce = "laitue"
-    case honey = "miel"
-    case hive = "ruche"
-    case raspberry = "framboise"
-    case grapefruit = "pamplemousse"
-    case thyme = "thym"
-    case olive = "olive"
-    case apricot = "abricot"
-    case grape = "raisin"
-    case tomato = "tomate"
-    case vegetable = "légumes frais"
-    case apple = "pomme"
-    case wheat = "blé"
-    case peach = "pêche"
-    case pear = "poire"
-    case celery = "céleris branche"
-    case corn = "maïs"
-    case watermelon = "pastèque"
-    case eggplant = "aubergine"
-    case zucchini = "courgette"
-    case pepper = "poivron"
-    case oliveOil = "huile d'olive"
-    case rawOliveOil = "huile d'olive, brute"
-    case potatoes = "pommes de terre"
-    case preserves = "conserves"
-    case carrot = "carotte"
-    case jam = "confiture"
-    case hemp = "chanvre"
-    case aromaticPlants = "épice"
-    case pig = "porc"
-    case meat = "viande"
-    case fig = "figue"
-    case kiwi = "kiwi"
-    case beet = "betterave"
-    case bread = "pain"
-    case rusk = "biscotte"
-    case gingerbread = "pains d'épices"
-    case sandwich = "sandwich"
-    case pastrie = "viennoiserie"
-}
